middleware/requirements.txt
appdirs==1.4.3
attrs==18.2.0
black==18.9b0
certifi==2018.4.16
chardet==3.0.4
click==6.7
flake8==3.6.0
Flask==1.0.2
Flask-Cors==3.0.6
gitdb2==2.0.5
GitPython==2.1.11
gunicorn==19.8.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
pbr==5.1.1
pycodestyle==2.4.0
pyflakes==2.0.0
python-dotenv==0.9.1
requests==2.19.1
six==1.11.0
smmap2==2.0.5
toml==0.10.0
urllib3==1.23
Werkzeug==0.14.1
